Navbar folder structure updated
