+++ src/components/Layouts/DashboardLayout/Navbar/NavbarDropdownItem.tsx
@@ -1,0 +1,15 @@
+import React, { ComponentPropsWithoutRef } from "react";
+
+function NavbarDropdownItem(props: ComponentPropsWithoutRef<"button">) {
+	return (
+		<button
+			{...props}
+			type="button"
+			className={`hover:bg-[--navbar-dropdown_menu_item-background-hover-color] flex justify-center items-center p-2 text-sm ${props.className}`}
+		>
+			{props.children}
+		</button>
+	);
+}
+
+export default NavbarDropdownItem;
